feat: make the base interface better. Many small change

+++ include/__impl/utils/bit_operation.hpp
@@ -10,12 +10,12 @@ namespace cpparmc::bits {
 
     template<typename T>
     inline void set_nth_bit(T& val, bool bit, std::size_t n) {
-        val = val | (bit << n);
+        val = val | (static_cast<std::uint64_t>(bit) << n);
     }
 
     template<typename T>
     inline void append_bit(T& origin, bool bit) {
-        origin = (origin << 1U) | bit;
+        origin = (origin << 1U) | static_cast<std::uint64_t>(bit);
     }
 
     inline std::uint64_t get_n_repeat_bit(bool bit, std::size_t n) {
@@ -32,12 +32,12 @@ namespace cpparmc::bits {
         if (width <= head) {
             const T buf = (origin & get_n_repeat_bit(true, head)) << (head - width);
             origin = 0U;
-            return { buf, 0U };
+            return {buf, 0U};
         } else {
             const auto rest_width = width - head;
             const T buf = (origin >> rest_width) & get_n_repeat_bit(true, head);
             origin &= get_n_repeat_bit(true, rest_width);
-            return { buf, rest_width };
+            return {buf, rest_width};
         }
     }
 }
